requirements.txt
aiohttp==3.10.5
python-dotenv==1.0.1
pydantic==2.9.1
pydantic-settings==2.0.3
pytest==8.3.2
pytest-asyncio==0.24.0
aresponses==3.0.0
sqlalchemy==2.0.34
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.8
fakeredis==2.24.1
alembic==1.13.2
pyyaml==6.0.1
coloredlogs==15.0.1
spacy==3.7.2
rapidfuzz==3.9.6
scikit-learn==1.5.2
pandas==2.2.2
statsmodels==0.14.2
py-clob-client>=0.21.0
